fix: libft flags and strichar

+++ libft/src/strichar.c
@@ -6,7 +6,7 @@
 /*   By: gcosta-d <gcosta-d@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/31 22:04:06 by gcosta-d          #+#    #+#             */
-/*   Updated: 2022/03/31 22:18:09 by gcosta-d         ###   ########.fr       */
+/*   Updated: 2022/04/27 21:08:19 by gcosta-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ int	strichar(char *string, int pos, char c)
 		return (-1);
 	while (string[pos] != c && string[pos])
 		pos++;
-	if (pos == ft_strlen(string))
+	if ((size_t)pos == ft_strlen(string))
 		return (-1);
 	return (pos);
 }
